fix(block): change hardcoded variables into dictionary
